requirements-dev.txt
attrs~=23.2
coverage~=7.4
mypy~=1.9
pytest-mock~=3.14
pytest~=8.1
ruff~=0.3
